Fix: Skin: No tab selected after re-creating.
-Fix: When using a skin, no tab would be selected upon re-creating the
Trades GUI from the Settings menu
This was due to the tabID (blank) value would be assigned to
currentActiveTab.
Added a check: If the tabID is not set, leave the currentActiveTab value
as it is.

Added a SplashText to notify the user that the GUI is re-creating
itself.

Changed the Support tickbox text to a more fitting one.