Code-wise changes + Small fixes
- Fixed a typo issue in Disable_Hotkeys func: indzex > index
- Fixed: Skin: Hovering a button, then hovering the background, and
going back to the same button would not switch to the "hovered" asset.
This was because the variable used to compare the current hovered button
and the previously hovered button would only be reseted upon hovering a
different control. It is now reset as soon as no control is hovered.

- Effectively removed the GlobalValues array.
- Created two new array to replace GlobalValues: TradesGui_Values and
ProgramSettings.
TradesGui_Values contains everything related to the Trades GUI, such as
the current active tab, the GUI handler, the height and width, ...
ProgramSettings contains the local ini file settings.
- Changes to many sub-array values.
- Renamed GameValues to GameSettings