1.10.7 - Dpi Factor fix
- The user can now enable Debug mode by creating a file named
"ENABLE_DEBUG.txt" in the same directory as the script. Debug options
are limited to a set of fake trade requests. When this setting is
enabled, trade infos will not be saved into the Trades_History.ini file.

- Gui_Trades_Set_Height(): The height is now multiplied by the screen
dpi. This fixes the issue where the GUI size would be incorrect, due to
the WinMove command not taking the screen dpi in account.
- A new command-line parameter has been added: Screen_DPI. It is used
internally by the script to keep the current user's DPI when reloading
with the admin account.