Re-included the old updater
The new updater does not work for pre-1.11 versions.
This is because older versions were relying on an older version of the updater that contained all the variables internally. The new version uses commandline parameters to pass variables from the main script to the updater.

For this reason, I have to re-include the old updater with its original name and rename the new one. Core script for 1.11 and above will have to use the new version of the updater.