Using GitHub Releases API
- Check_Update() now makes use of the GitHub API.
You can only check for update once every 35mins. This is to prevent from spamming the API when restarting the app. This timer can be bypassed by the new button in the About GUI: Check Update.
In case we fail to reach the API, we still use the version.txt method as a fallback solution.
- Increased the timer for Check_Update() to 140mins (from 30).
- Added a button to check for update in the About GUI.
This button will not automatically run the download if a  different version is available.
- On start, if the version contains BETA, the Update_Beta value is set to 1 in the ini file. This is now also true when the user starts with a non-beta version, with a value set to 0 in this case.