Typo fix
- Wrong var names used.
Normally, on launch, to avoid spamming Github API, we only check for update if the latest check is older than 35mins. After checking for update, we write the latest stable and beta tag to the ini file.
In case the check is not older than 35mins, we read the ini file to retrieve the version tags. This allows to show the "update avialable" notification on startup without having to contact the GitHub API again.