Actions list changes
- Updated the actions list with a more intuitive actions-pushing system.
The three "Replace se..", "Push at end/se" buttons have been removed and replaced with two new ones:

- - "Save changes to action..."
Shows a menu that allows to save the changes to a specific action number.
This number can be the currently selected action, or any other one.
- - "Add as a new action"
Allows to push the action type and content as a new action. This action will be added to the end of the list.
In case the button's last action is "Close tab" or "Write Message", the new action will be pushed as previous to last.

- Actions can now be re-ordered by right clicking on them and using "Move Up" or "Move Down".
- Minor re-ordering of "Specials" section action list.
- Fix: Pressing up arrow when being on top of action list now correctly select the first action, instead of selected the sectio name "--- example ---"
- Fix: Selecting an empty action name now select the next available action